pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
description = "Fully-local LiveKit voice agent (Ollama LLM + Faster-Whisper STT + local TTS endpoint)"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  # LiveKit Agents core + only the plugins you need (OpenAI-compatible for Ollama/STT/TTS, Silero VAD, Turn Detector)
  "livekit-agents[openai,silero,turn-detector,mcp]~=1.3.9",
  "mcp>=1.25.0",

  # Faster-Whisper STT (CTranslate2 GPU on Windows requires CUDA 12.x installed)
  "faster-whisper>=1.1.0",
  "ctranslate2>=4.6.0",
  "numpy>=2.0.0",

  # Local OpenAI-compatible STT microservice (FastAPI)
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",

  "python-dotenv>=1.0.1",
  "pyperclip>=1.9.0",
  "pywin32==311",
  "psutil==7.2.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.25.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.9.0",
  "mypy>=1.10.0",
  "black>=24.0.0",
]
gpu = [
  # CUDA runtime DLLs (cu12 = CUDA 12 family)
  "nvidia-cuda-runtime-cu12==12.9.79; platform_system=='Windows' and platform_machine=='AMD64'",

  # cuBLAS runtime DLLs
  "nvidia-cublas-cu12==12.9.1.4; platform_system=='Windows' and platform_machine=='AMD64'",

  # cuDNN runtime DLLs
  "nvidia-cudnn-cu12==9.17.1.4; platform_system=='Windows' and platform_machine=='AMD64'",

  # NVRTC runtime wheels (often referenced in Faster-Whisper troubleshooting)
  "nvidia-cuda-nvrtc-cu12>=12.4.127; platform_system=='Windows' and platform_machine=='AMD64'",
]


[tool.black]
line-length = 100
target-version = ["py311", "py312", "py313"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
